Windows: Refactor container
