+++ daemon/volumes_windows.go
@@ -6,3 +6,11 @@ package daemon
 func copyOwnership(source, destination string) error {
 	return nil
 }
+
+func (container *Container) prepareVolumes() error {
+	return nil
+}
+
+func (container *Container) setupMounts() error {
+	return nil
+}
